break out book_report app into users and reports modules. Nullable username.
