--- book_report/tests.py
@@ -1,8 +1,0 @@
-import pytest
-
-
-def test_test_file_exists():
-    pass
-
-
-# Create your tests here.
