+++ book_report/users/models.py
@@ -5,7 +5,7 @@ from django.utils import timezone
 class User(models.Model):
     goodreads_user_id = models.IntegerField()
     name = models.CharField(max_length=200)
-    username = models.CharField(max_length=200)
+    username = models.CharField(max_length=200, null=True)
     visibility = models.BooleanField()
     access_token = models.CharField(max_length=255)
     access_secret = models.CharField(max_length=255)
