+++ book_report/users/views.py
@@ -9,8 +9,8 @@ from django.shortcuts import render, redirect
 from goodreads.client import GoodreadsClient
 from goodreads.session import GoodreadsSession
 
-
-from book_report.models import User
+from book_report.users.models import User
+from book_report.utils import user_is_visible
 
 gc = GoodreadsClient(os.environ["GOODREADS_KEY"], os.environ["GOODREADS_SECRET"])
 
@@ -57,7 +57,7 @@ def get_or_create_user(goodreads_user):
             username=goodreads_user.user_name,
             access_token=gc.session.access_token,
             access_secret=gc.session.access_token_secret,
-            visibility=0,
+            visibility=1,
             created=datetime.datetime.utcnow(),
             updated=datetime.datetime.utcnow(),
         )
@@ -93,56 +93,3 @@ def get_all_shelf_reviews_for_user(request, user_gid, shelf_name="read"):
     reviews = gc.user(user_gid).per_shelf_reviews(shelf_name=shelf_name)
 
     return render(request, "book_report/user_shelf_reviews.html", {"reviews": reviews})
-
-
-def identify_controversial_opinions(request, user_gid, report_length=20):
-    """
-    Note: ratings with a review score of '0' will be discarded as they skew the results to books on your read shelf that haven't been rated.
-    :param request:
-    :param user_gid:
-    :param report_length:
-    :return:
-    """
-    if not user_is_visible(request.session.get("gid"), user_gid):
-        raise Http404("User not found")
-
-    reviews = gc.user(user_gid).per_shelf_reviews(shelf_name="read")
-
-    diffed_reviews = []
-
-    for review in reviews:
-        if review.rating == "0":
-            continue
-        average_score = review.book["average_rating"]
-        diff = abs(float(average_score) - float(review.rating))
-        diffed_reviews.append((diff, review))
-
-    diffed_reviews.sort(key=operator.itemgetter(0), reverse=True)
-    print(diffed_reviews)
-    return render(
-        request,
-        "book_report/user_opinions_report.html",
-        {"reviews": [r[1] for r in diffed_reviews][:report_length]},
-    )
-
-
-def user_is_visible(session_user_id, user_gid):
-    """
-    Helper function that determines if a requested user is visible to the logged in user.
-    :param session_user_id:
-    :param user_gid:
-    :return:
-    """
-    if user_gid == session_user_id:
-        visible = True
-    else:
-        try:
-            user = User.objects.get(goodreads_user_id=user_gid)
-            if user.visibility == 0:
-                visible = False
-            else:
-                visible = True
-        except ObjectDoesNotExist:
-            visible = False
-
-    return visible
