+++ bookpinions/urls.py
@@ -13,7 +13,9 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
 from django.urls import path, include
 
-urlpatterns = [path("admin/", admin.site.urls), path("", include("book_report.urls"))]
+urlpatterns = [
+    path("users/", include("book_report.users.urls")),
+    path("report/", include("book_report.reports.urls")),
+]
